Improve typing so that TMCM3110 example types smoothly
- Hoist __enter__/__exit__ from subclasses into `TmclInterface`
- Add a few `if X is not None`
- Add config for consistent import sorting
