+++ pyproject.toml
@@ -38,3 +38,9 @@ find = {}
 
 [tool.setuptools.dynamic]
 version = { attr = "pytrinamic.version.__version__" }
+
+[tool.ruff.lint.isort]
+force-single-line = true
+
+[tool.isort]
+force_single_line = true
