+++ pytrinamic/connections/dummy_tmcl_interface.py
@@ -26,22 +26,6 @@ class DummyTmclInterface(TmclInterface):
 
         self.logger.debug("Opening port (baudrate=%s).", datarate)
 
-    def __enter__(self):
-        return self
-
-    def __exit__(self, exit_type, value, traceback):
-        """
-        Close the connection at the end of a with-statement block.
-        """
-        del exit_type, value, traceback
-        self.close()
-
-    def close(self):
-        """
-        Closes the dummy TMCL connection
-        """
-
-
     def _send(self, host_id, module_id, data):
         """
             Send the bytearray parameter [data].
